+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [3.13.0](https://github.com/coinranking/exchanges/compare/v3.12.0...v3.13.0) (2020-08-11)
+
+
+### Features
+
+* **polarity:** use the waves driver ([5f7a7ec](https://github.com/coinranking/exchanges/commit/5f7a7ec19a49f478464ae7101d7668d845bcfb83))
+* polarity driver ([ebd0ba0](https://github.com/coinranking/exchanges/commit/ebd0ba09555588f671273acdde55bd662ea541c3))
+
+# [3.12.0](https://github.com/coinranking/exchanges/compare/v3.11.0...v3.12.0) (2020-08-11)
+
+
+### Features
+
+* ankerswap driver ([68549ab](https://github.com/coinranking/exchanges/commit/68549ab33da852e5a0e1dc4a371c5d63f10e3c51))
+* bololex driver ([4df8bb8](https://github.com/coinranking/exchanges/commit/4df8bb873ffd85933874a98e7883a3c9bdb5a5a7))
+
 # [3.11.0](https://github.com/coinranking/exchanges/compare/v3.10.0...v3.11.0) (2020-07-30)